(refs #483) Rename handle-changes.js to handle-pouch-changes.js
